refactor: Separate the Monaco Editor into it's own component so it can be loaded dynamically
