+++ src/aux-server/aux-web/shared/vue-components/SimpleTagEditor/SimpleTagEditor.css
@@ -1,0 +1,20 @@
+.simple-editor.formula {
+    font-family: Consolas, 'Courier New', monospace;
+    letter-spacing: 0;
+}
+
+.simple-editor {
+    width: 100%;
+    max-height: 400px;
+    white-space: nowrap;
+    overflow-wrap: normal;
+    overflow-x: auto;
+}
+
+.simple-editor-wrapper {
+    padding: 5px;
+}
+
+.simple-editor-wrapper > .md-field {
+    margin-bottom: 0;
+}
