--- src/aux-server/aux-web/shared/vue-components/TagValueEditor/TagValueEditor.css
@@ -1,42 +1,0 @@
-.simple-editor.formula {
-    font-family: Consolas, 'Courier New', monospace;
-    letter-spacing: 0;
-}
-
-.simple-editor {
-    width: 100%;
-    max-height: 400px;
-    white-space: nowrap;
-    overflow-wrap: normal;
-    overflow-x: auto;
-}
-
-.simple-editor-wrapper {
-    padding: 5px;
-}
-
-.simple-editor-wrapper > .md-field {
-    margin-bottom: 0;
-}
-
-.editor-wrapper {
-    width: 100%;
-    height: calc(100% - 8px);
-    border-top: solid 1px;
-    margin-top: 8px;
-}
-
-.code-editor-wrapper {
-    width: 100%;
-
-    /* Height is 100% the container minus the height of the breadcrumbs container */
-    height: calc(100% - 23px);
-}
-
-.editor-breadcrumbs {
-    width: 100%;
-    padding: 0 14px;
-    line-height: 22px;
-    background-color: var(--disabled-cell-color);
-    border-bottom: solid 1px var(--file-cell-border-color);
-}
