+++ src/aux-server/aux-web/shared/vue-components/VueHelpers.ts
@@ -1,0 +1,9 @@
+/**
+ * Determines if the given element is focused.
+ */
+export function isFocused(el: HTMLElement) {
+    if (el && document.activeElement) {
+        return el.contains(document.activeElement);
+    }
+    return false;
+}
